Datepicker is separated in audit log
