+++ apps/web-giddh/src/app/audit-logs/audit-logs.component.html
@@ -1,9 +1,15 @@
 <div class="container-fluid">
+    <div class="custom-datepicker" (click)="showGiddhDatepicker($event)" *ngIf="isNewVersion">
+        <div>
+            <i><img src="assets/images/custom-calender.svg"></i>
+            <input type="text" name="selectedDateRange" [value]="(selectedDateRangeUi) ? selectedDateRangeUi : ''" class="giddh-datepicker date-range-picker" />
+        </div>
+    </div>
     <div class="row">
         <div class="col-xs-3 max350">
             <!--sidebar-->
             <audit-logs-sidebar *ngIf="!isNewVersion"></audit-logs-sidebar>
-            <audit-logs-form *ngIf="isNewVersion"></audit-logs-form>
+            <audit-logs-form *ngIf="isNewVersion" [fromDate]="fromDate" [toDate]="toDate"></audit-logs-form>
             <!--/sidebar-->
         </div>
         <div class="col-xs-9">
@@ -12,3 +18,13 @@
         </div>
     </div>
 </div>
+
+<ng-template #datepickerTemplate>
+    <div class="datepicker-modal">
+        <div class="modal-body">
+            <app-datepicker-wrapper [inputStartDate]="selectedDateRange?.startDate" [inputEndDate]="selectedDateRange?.endDate" [alwaysShowCalendars]="true" [ranges]="datePickerOptions.ranges" [selectedRangeLabel]="selectedRangeLabel" [showCustomRangeLabel]="true"
+                [showClearButton]="false" [showCancel]="true" [linkedCalendars]="true" [showDropdowns]="true" [locale]="{applyLabel: 'Done'}" (rangeClicked)="dateSelectedCallback($event)" (datesUpdated)="dateSelectedCallback($event)" [keepCalendarOpeningWithRange]="false"
+                [showRangeLabelOnInput]="false" [dateFieldPosition]="dateFieldPosition"></app-datepicker-wrapper>
+        </div>
+    </div>
+</ng-template>
